made sure to fall back to blocking backup in case of no WAL

+++ distributed/src/main/java/com/orientechnologies/orient/server/distributed/impl/task/OSyncDatabaseTask.java
@@ -140,6 +140,8 @@ public class OSyncDatabaseTask extends OAbstractSyncDatabaseTask {
                   try {
                     OWriteAheadLog wal = ((OAbstractPaginatedStorage) database.getStorage().getUnderlying()).getWALInstance();
                     OLogSequenceNumber lsn = wal.end();
+                    if (lsn == null)
+                      throw new UnsupportedOperationException();
                     wal.addCutTillLimit(lsn);
 
                     resultedBackupFile.delete();
